fixed bugs of database

+++ new_MyApplication/MyApplication/app/src/main/java/com/example/huiyi/myapplication/StatusDao.java
@@ -16,5 +16,5 @@ public interface StatusDao {
     public void updateStatus(Status status);
 
     @Query("SELECT * FROM status WHERE status.sid = :sid")
-    List<Status> getStatusById(int sid);
+    List<Status> getStatusById(long sid);
 }
